PyNCBIminer_1.2.11_MacOS_source_code/requirements.txt
altgraph==0.17.3
biopython==1.79
certifi==2024.2.2
chardet==3.0.4
cycler==0.11.0
fonttools==4.40.0
func-timeout==4.3.5
idna==2.7
joblib==1.3.1
kiwisolver==1.4.4
macholib==1.16.2
markov-clustering==0.0.6.dev0
matplotlib==3.5.1
networkx==2.8.4
numpy==1.21.2
packaging==23.1
pandas==1.3.4
Pillow==10.0.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2023.5
pyparsing==3.1.0
PySide2==5.15.2.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.19.1
scikit-learn==1.3.0
scipy==1.8.1
shiboken2==5.15.2.1
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.23
